build: fix: compose process ignores TERM signal when resolving deps
On big docker compose projects with a lot of dependencies, the
docker-compose plugin process does not terminate when the process group
receives a TERM or INT signal.
The docker-compose process continues to run and does not react to any
TERM signals.

When a TERM signal is received, compose cancels the context.
The code that resolves the dependency tree is not checking if the
context gets canceled and does not abort.
Check if the context is canceled when receiving or sending to NodeCh
and terminate in that case.

With this change, docker-compose returns immediately when an INT or TERM
signal is received, also on the big compose project where it got stuck
before.